genai/requirements.txt
# For FastAPI web framework and server
fastapi
uvicorn[standard]
pydantic

# For LangChain and OpenAI models
langchain
langchain-openai
openai

# For loading environment variables
python-dotenv

# For testing
pytest
requests
httpx